requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-environ==0.10.0
django-filter==23.2
Pillow==9.5.0
psycopg2-binary==2.9.6
redis==4.5.5
celery==5.3.0
requests==2.31.0
pandas==2.0.2
numpy==1.24.3
scikit-learn==1.2.2
scipy==1.10.1
statsmodels==0.13.5
tensorflow==2.12.0
plotly==5.14.0
python-dotenv==1.0.0
gunicorn==20.1.0
whitenoise==6.4.0
drf-spectacular==0.26.1
pytest==7.3.2
pytest-django==4.5.2
pytest-cov==4.1.0
black==23.3.0
flake8==6.0.0
isort==5.12.0
